+++ src/testbed.h
@@ -6,7 +6,7 @@
 #include <gl/GL.h>
 
 void testbed_init(void);
-void testbed_update(GLsizei width, GLsizei height, long double time, long double delta_time);
+void testbed_update(GLsizei width, GLsizei height, float time, float delta_time);
 void testbed_reload(void);
 
 #endif // SHADER_TESTBED_TESTBED_H_
